requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
sqlalchemy==2.0.48
python-dotenv==1.2.1
httpx==0.28.1
google-genai==1.47.0
apscheduler==3.11.2
pytest==8.4.2
pydantic==2.12.5
